x-bot/requirements.txt
tweepy==4.14.0
google-genai>=1.0.0
schedule==1.2.2
python-dotenv==1.0.1
